stats/requirements.txt
Django==5.0.4
Psycopg2==2.9.9
gunicorn==22.0.0
#python-dotenv==1.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.2.2
PyJWT==2.8.0
pyotp==2.9.0
pillow==10.3.0